requirements.txt
Flask==3.0.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36
alembic==1.13.2
python-dotenv==1.0.1
boto3==1.34.162
jinja2==3.1.4
werkzeug==3.0.3